+++ test/domain/usecases/movie/get_movie_recommendations_test.dart
@@ -4,7 +4,7 @@ import 'package:ditonton/domain/usecases/movie/get_movie_recommendations.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../../helpers/movie_test_helper.mocks.dart';
+import '../../../helpers/movie_test_helper.mocks.dart';
 
 void main() {
   late GetMovieRecommendations usecase;
