+++ test/domain/usecases/movie/get_now_playing_movies_test.dart
@@ -1,25 +1,25 @@
 import 'package:dartz/dartz.dart';
 import 'package:ditonton/domain/entities/movie/movie.dart';
-import 'package:ditonton/domain/usecases/movie/get_top_rated_movies.dart';
+import 'package:ditonton/domain/usecases/movie/get_now_playing_movies.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../../helpers/movie_test_helper.mocks.dart';
+import '../../../helpers/movie_test_helper.mocks.dart';
 
 void main() {
-  late GetTopRatedMovies usecase;
+  late GetNowPlayingMovies usecase;
   late MockMovieRepository mockMovieRepository;
 
   setUp(() {
     mockMovieRepository = MockMovieRepository();
-    usecase = GetTopRatedMovies(mockMovieRepository);
+    usecase = GetNowPlayingMovies(mockMovieRepository);
   });
 
   final tMovies = <Movie>[];
 
-  test('should get list of movies from repository', () async {
+  test('should get list of movies from the repository', () async {
     // arrange
-    when(mockMovieRepository.getTopRatedMovies())
+    when(mockMovieRepository.getNowPlayingMovies())
         .thenAnswer((_) async => Right(tMovies));
     // act
     final result = await usecase.execute();
