+++ test/domain/usecases/movie/get_popular_movies_test.dart
@@ -4,7 +4,7 @@ import 'package:ditonton/domain/usecases/movie/get_popular_movies.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../../helpers/movie_test_helper.mocks.dart';
+import '../../../helpers/movie_test_helper.mocks.dart';
 
 void main() {
   late GetPopularMovies usecase;
