+++ test/domain/usecases/movie/get_top_rated_movies_test.dart
@@ -1,25 +1,25 @@
 import 'package:dartz/dartz.dart';
 import 'package:ditonton/domain/entities/movie/movie.dart';
-import 'package:ditonton/domain/usecases/movie/get_now_playing_movies.dart';
+import 'package:ditonton/domain/usecases/movie/get_top_rated_movies.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../../helpers/movie_test_helper.mocks.dart';
+import '../../../helpers/movie_test_helper.mocks.dart';
 
 void main() {
-  late GetNowPlayingMovies usecase;
+  late GetTopRatedMovies usecase;
   late MockMovieRepository mockMovieRepository;
 
   setUp(() {
     mockMovieRepository = MockMovieRepository();
-    usecase = GetNowPlayingMovies(mockMovieRepository);
+    usecase = GetTopRatedMovies(mockMovieRepository);
   });
 
   final tMovies = <Movie>[];
 
-  test('should get list of movies from the repository', () async {
+  test('should get list of movies from repository', () async {
     // arrange
-    when(mockMovieRepository.getNowPlayingMovies())
+    when(mockMovieRepository.getTopRatedMovies())
         .thenAnswer((_) async => Right(tMovies));
     // act
     final result = await usecase.execute();
