+++ test/domain/usecases/movie/get_watchlist_status_test.dart
@@ -2,7 +2,7 @@ import 'package:ditonton/domain/usecases/movie/get_watchlist_status.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../../helpers/movie_test_helper.mocks.dart';
+import '../../../helpers/movie_test_helper.mocks.dart';
 
 void main() {
   late GetWatchListStatus usecase;
