+++ test/domain/usecases/movie/search_movies_test.dart
@@ -4,7 +4,7 @@ import 'package:ditonton/domain/usecases/movie/search_movies.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
-import '../../helpers/movie_test_helper.mocks.dart';
+import '../../../helpers/movie_test_helper.mocks.dart';
 
 void main() {
   late SearchMovies usecase;
